+++ core/src/main/java/de/fraunhofer/ids/messaging/core/daps/TokenManagerService.java
@@ -25,12 +25,10 @@ public interface TokenManagerService {
      * @return The DAT as String
      * @throws DapsConnectionException Thrown if no connection to DAPS possible
      * @throws DapsEmptyResponseException Thrown if DAPS returned invalid response
-     * @throws ConnectorMissingCertExtensionException Thrown if something is
      * wrong with the Connector Certificate
      */
     String acquireToken(String dapsURL)
             throws
             DapsConnectionException,
-            DapsEmptyResponseException,
-            ConnectorMissingCertExtensionException;
+            DapsEmptyResponseException;
 }